rpc: move ValidateAddress to the result package
Because it is a result of the RPC call.
