+++ pkg/rpc/result/validate_address.go
@@ -1,9 +1,9 @@
-package wrappers
+package result
 
-// ValidateAddressResponse represents response to validate address call. Notice
+// ValidateAddress represents result of the `validateaddress` call. Notice that
 // Address is an interface{} here because server echoes back whatever address
 // value user has sent to it, even if it's not a string.
-type ValidateAddressResponse struct {
+type ValidateAddress struct {
 	Address interface{} `json:"address"`
 	IsValid bool        `json:"isvalid"`
 }
